requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
pyyaml==6.0.2
python-multipart==0.0.20
starlette==0.41.3
anyio==4.8.0
click==8.1.8
cryptography==44.0.0
